refactoring into objects
