+++ core/src/com/glazdans/echo/bullet/GameObject.java
@@ -88,7 +88,7 @@ public class GameObject {
         direction = direction.mul(rotation);
         Vector3 startingPosition = new Vector3(direction);
         startingPosition.scl(1f).add(position);
-        Projectile projectile = new Projectile(startingPosition,direction,21f,BulletTestScreen.collisionWorld);
+        Projectile projectile = new Projectile(startingPosition,direction,21f,Physics.getInstance().collisionWorld);
         BulletTestScreen.projectiles.add(projectile);
     }
